backend/cloudFunctions/leadProcessor/requirements.txt
functions-framework==3.*
flask==2.*
google-cloud-firestore==2.*
google-generativeai>=0.3.1
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
requests>=2.28.2
